패치 내역 분리
